requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
Pillow==10.4.0
tiktoken==0.7.0
openai==1.48.0
httpx==0.24.0